requirements.txt
dagshub == 0.3.8
mlflow == 2.7.1
dvc == 3.24.0
matplotlib == 3.5.2
rdkit == 2023.3.3
scikit-learn == 0.24.2
numpy == 1.22.3
pandas == 1.4.3
seaborn == 0.13.0

hydra-core == 1.3.2
omegaconf == 2.1.2

wandb == 0.13.1
pyopenssl == 22.1.0
shap == 0.43.0

# deepchem[torch] == 2.7.1
# torch_geometric == 2.3.1
# molfeat == 0.9.4
# datamol == 0.11.4
# chemplot == *

## contained in MegaMolBART v0.2 docker container:
# nemo-chem                     0.2.0.dev0
# nemo-toolkit                  1.11.0
# pytorch-lightning             1.6.5
# torch                         1.13.0a0+08820cb
